🚚 Rename:  cssColorKeywords.ts 파일 utils -> constants 폴더로 이동
